requirements.txt
absl-py==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
appdirs==1.4.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.3.0
attrs==26.1.0
audio-metrics==1.0.4
audioread==3.1.0
babel==2.18.0
beautifulsoup4==4.14.3
bleach==6.3.0
braceexpand==0.1.7
cdpam==0.0.6
certifi==2026.4.22
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.7
click==8.3.3
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
coverage==7.13.5
cycler==0.12.1
cython==3.2.4
datasets==2.21.0
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
dill==0.3.8
distlib==0.4.0
einops==0.8.2
executing==2.2.1
fastdtw==0.3.4
fastjsonschema==2.21.2
ffmpeg-python==0.2.0
filelock==3.29.0
flatbuffers==25.12.19
fonttools==4.62.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2024.6.1
ftfy==6.3.1
future==1.0.0
gammatone==1.0.3
ghp-import==2.1.0
gitdb==4.0.12
gitpython==3.1.49
grpcio==1.80.0
h11==0.16.0
h5py==3.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.12.2
identify==2.6.19
idna==3.13
iniconfig==2.3.0
ipykernel==7.2.0
ipython==9.13.0
ipython-pygments-lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
jinja2==3.1.6
joblib==1.5.3
json5==0.14.0
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter-client==8.8.0
jupyter-core==5.9.1
jupyter-events==0.12.1
jupyter-lsp==2.3.1
jupyter-server==2.17.0
jupyter-server-terminals==0.5.4
jupyterlab==4.5.7
jupyterlab-pygments==0.3.0
jupyterlab-server==2.28.0
kiwisolver==1.5.0
laion-clap==1.1.7
lark==1.3.1
lazy-loader==0.5
librosa==0.11.0
lightning==2.6.1
lightning-utilities==0.15.3
llvmlite==0.47.0
loguru==0.7.3
markdown==3.10.2
markdown-it-py==4.0.0
markupsafe==3.0.3
matplotlib==3.10.9
matplotlib-inline==0.2.1
mdurl==0.1.2
mergedeep==1.3.4
mistune==3.2.0
mkdocs==1.6.1
mkdocs-get-deps==0.2.2
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
multiprocess==0.70.16
musdb==0.4.3
nbclient==0.10.4
nbconvert==7.17.1
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
nodeenv==1.10.0
notebook==7.5.5
notebook-shim==0.2.4
numba==0.65.1
numpy==1.26.4
numpy-audio-limiter==0.2.0
onnxruntime==1.25.1
opt-einsum==3.4.0
packaging==26.2
pandas==3.0.2
pandocfilters==1.5.1
parso==0.8.6
pathspec==1.1.1
pesq==0.0.4
pillow==12.2.0
pip==26.1
platformdirs==4.9.6
pluggy==1.6.0
pooch==1.9.0
prdc==0.2
pre-commit==4.6.0
progressbar==2.5
prometheus-client==0.25.0
prompt-toolkit==3.0.52
propcache==0.4.1
protobuf==7.34.1
psutil==7.2.2
pure-eval==0.2.3
pyaml==26.2.1
pyarrow==24.0.0
pycparser==3.0
pydantic==2.13.3
pydantic-core==2.46.3
pygments==2.20.0
pyloudnorm==0.2.0
pymcd==0.2.1
pyparsing==3.3.2
pysptk==1.0.1
pystoi==0.4.1
pytest==9.0.3
python-dateutil==2.9.0.post0
python-discovery==1.2.2
python-dotenv==1.2.2
python-json-logger==4.1.0
pytorch-lightning==2.6.1
pyworld==0.3.5
pyyaml==6.0.3
pyyaml-env-tag==1.1
pyzmq==27.1.0
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
resampy==0.4.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==15.0.0
rpds-py==0.30.0
ruff==0.15.12
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
send2trash==2.1.0
sentry-sdk==2.58.0
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
soundfile==0.13.1
soupsieve==2.8.3
soxr==1.0.0
stack-data==0.6.3
stempeg==0.2.6
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.22.2
torchlibrosa==0.1.0
torchmetrics==1.9.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
transformers==5.7.0
ty==0.0.33
typer==0.25.0
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2026.2
uri-template==1.3.0
urllib3==2.6.3
virtualenv==21.3.0
wandb==0.26.1
watchdog==6.0.0
wcwidth==0.6.0
webcolors==25.10.0
webdataset==1.0.2
webencodings==0.5.1
websocket-client==1.9.0
werkzeug==3.1.8
wget==3.2
xxhash==3.7.0
yarl==1.23.0
yt-dlp==2026.3.17
zimtohrli==0.2.1.post1
